# API-Gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AchievementCreateResponse:
    description: Ответ с данными для загрузки файла достижения
    properties:
      meta:
        $ref: '#/definitions/models.AchievementMeta'
      upload_url:
        $ref: '#/definitions/models.UploadUrlResponse'
    type: object
  models.AchievementList:
    description: Список достижений пользователя
    properties:
      achievements:
        items:
          $ref: '#/definitions/models.AchievementMeta'
        type: array
    type: object
  models.AchievementMeta:
    description: Метаданные достижения пользователя
    properties:
      created_at:
        example: "2023-10-05T14:30:00Z"
        type: string
      file_name:
        example: diploma.pdf
        type: string
      file_size:
        example: 2048576
        type: integer
      file_type:
        example: application/pdf
        type: string
      name:
        example: Диплом бакалавра
        type: string
      user_uuid:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  models.AchievementUploadRequest:
    description: Запрос на подготовку загрузки достижения
    properties:
      file_name:
        example: diploma.pdf
        type: string
      file_size:
        example: 2048576
        minimum: 1
        type: integer
      file_type:
        example: application/pdf
        type: string
      name:
        example: Диплом бакалавра
        type: string
    required:
    - file_name
    - file_size
    - file_type
    - name
    type: object
  models.AchievementUrl:
    description: URL для скачивания файла достижения
    properties:
      expires_at:
        example: 1696523400
        type: integer
      url:
        example: https://s3.amazonaws.com/bucket/file.pdf?signature=...
        type: string
    type: object
...
  models.ErrorResponse:
    description: Расширенный ответ с ошибкой
    properties:
      code:
        example: INVALID_INPUT
        type: string
      details:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
...
  models.UploadUrlResponse:
    description: URL для загрузки файла в S3
    properties:
      expires_at:
        example: 1696523400
        type: integer
      s3_key:
        example: achievements/user123/diploma.pdf
        type: string
      upload_url:
        example: https://s3.amazonaws.com/bucket/upload?signature=...
        type: string
    type: object
...
  models.ValidationError:
    description: Детализированная ошибка валидации
    properties:
      code:
        example: REQUIRED_FIELD
        type: string
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@studjobs.ru
    name: API Support
  description: API Gateway для платформы StudJobs HH
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: StudJobs HH API Gateway
  version: "1.0"
paths:
...
  /user/achievements:
    get:
      consumes:
      - application/json
      description: Возвращает список всех достижений текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список достижений
          schema:
            $ref: '#/definitions/models.AchievementList'
        "400":
          description: Неверный ID пользователя
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить достижения пользователя
      tags:
      - Achievements
    post:
      consumes:
      - application/json
      description: Создает новое достижение и возвращает URL для загрузки файла
      parameters:
      - description: Данные достижения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AchievementUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные для загрузки файла
          schema:
            $ref: '#/definitions/models.AchievementCreateResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать достижение
      tags:
      - Achievements
  /user/achievements/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет достижение пользователя по названию
      parameters:
      - description: Название достижения
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Достижение не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить достижение
      tags:
      - Achievements
  /user/achievements/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Подтверждает успешную загрузку файла достижения и сохраняет метаданные
      parameters:
      - description: Название достижения
        in: path
        name: id
        required: true
        type: string
      - description: S3 ключ файла
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном подтверждении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Достижение не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Подтвердить загрузку достижения
      tags:
      - Achievements
  /user/achievements/{id}/download:
    get:
      consumes:
      - application/json
      description: Возвращает presigned URL для скачивания файла достижения
      parameters:
      - description: Название достижения
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL для скачивания
          schema:
            $ref: '#/definitions/models.AchievementUrl'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Достижение не найдено
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить URL для скачивания достижения
      tags:
      - Achievements
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RoleAuth:
    description: Роль пользователя
    in: header
    name: X-User-Role
    type: apiKey
swagger: "2.0"